Replaced DownloadFromZoneEffect with DownloadCardFromZoneAction
