+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/cards/blueprints/Blueprint212_019.java
@@ -1,10 +1,8 @@
 package com.gempukku.stccg.cards.blueprints;
 
 import com.gempukku.stccg.actions.Action;
-import com.gempukku.stccg.actions.Effect;
 import com.gempukku.stccg.actions.EffectResult;
-import com.gempukku.stccg.actions.PassThroughEffectAction;
-import com.gempukku.stccg.actions.playcard.ChooseAndPlayCardFromZoneEffect;
+import com.gempukku.stccg.actions.playcard.DownloadCardFromZoneAction;
 import com.gempukku.stccg.cards.blueprints.actionsource.SeedCardActionSource;
 import com.gempukku.stccg.cards.physicalcard.PhysicalCard;
 import com.gempukku.stccg.common.filterable.CardType;
@@ -31,9 +29,7 @@ public class Blueprint212_019 extends CardBlueprint {
     public List<Action> getValidResponses(PhysicalCard card, Player player, EffectResult effectResult) {
         List<Action> actions = new ArrayList<>();
         if (effectResult.getType() == EffectResult.Type.START_OF_MISSION_ATTEMPT) {
-            Effect chooseCardEffect = new ChooseAndPlayCardFromZoneEffect(Zone.DRAW_DECK, player, CardType.PERSONNEL);
-            Action action = new PassThroughEffectAction(card, chooseCardEffect);
-            actions.add(action);
+            actions.add(new DownloadCardFromZoneAction(Zone.DRAW_DECK, player, card, CardType.PERSONNEL));
         }
         return actions;
     }
